chore(program): Move CLI module under binary crate
Since the CLI logic is the main part of a binary crate, the
corresponding module `cli` is moved under the binary crate.
